chore: move slower tests to e2e folder
Introduces `npm run test:fast` to run faster tests.
